+++ src/redux/channel/index.ts
@@ -1,0 +1,6 @@
+export { join, leave, activate } from './action';
+export {
+  selectJoinLoading,
+  selectLeaveLoading,
+  selectJoined,
+} from './selector';