+++ src/redux/reducer.ts
@@ -1,0 +1,6 @@
+import channel from './channel/slice';
+import user from './user/slice';
+
+const reducer = { channel, user };
+
+export default reducer;